+++ language/move-binary-format/Cargo.toml
@@ -24,7 +24,9 @@ diem-workspace-hack = { version = "0.1", path = "../../crates/diem-workspace-hac
 [dev-dependencies]
 proptest = "1.0.0"
 proptest-derive = "0.3.0"
+move-bytecode-verifier = { path = "../move-bytecode-verifier" }
 move-core-types = { path = "../move-core/types", features = ["fuzzing"] }
+move-ir-to-bytecode = { path = "../move-ir-compiler/move-ir-to-bytecode" }
 serde_json = "1.0.64"
 
 [features]
