+++ language/move-ir-compiler/src/unit_tests/mod.rs
@@ -5,5 +5,4 @@
 #[macro_use]
 pub(crate) mod testutils;
 
-mod cfg_tests;
 mod function_tests;
