+++ language/move-ir-compiler/src/unit_tests/testutils.rs
@@ -4,45 +4,10 @@
 use anyhow::Result;
 use move_binary_format::{
     errors::{Location, VMError},
-    file_format::{CompiledModule, CompiledScript},
+    file_format::CompiledModule,
 };
-use move_bytecode_verifier::{verify_module, verify_script};
-use move_ir_to_bytecode::{
-    compiler::{compile_module, compile_script},
-    parser::{parse_module, parse_script},
-};
-
-fn compile_script_string_impl(
-    code: &str,
-    deps: Vec<CompiledModule>,
-) -> Result<(CompiledScript, Option<VMError>)> {
-    let parsed_script = parse_script(code).unwrap();
-    let script = compile_script(parsed_script, &deps)?.0;
-
-    let mut serialized_script = Vec::<u8>::new();
-    script.serialize(&mut serialized_script)?;
-    let deserialized_script = CompiledScript::deserialize(&serialized_script)
-        .map_err(|e| e.finish(Location::Undefined).into_vm_status())?;
-    assert_eq!(script, deserialized_script);
-
-    Ok(match verify_script(&script) {
-        Ok(_) => (script, None),
-        Err(error) => (script, Some(error)),
-    })
-}
-
-fn compile_script_string_and_assert_no_error(
-    code: &str,
-    deps: Vec<CompiledModule>,
-) -> Result<CompiledScript> {
-    let (verified_script, verification_error) = compile_script_string_impl(code, deps)?;
-    assert!(verification_error.is_none());
-    Ok(verified_script)
-}
-
-pub fn compile_script_string(code: &str) -> Result<CompiledScript> {
-    compile_script_string_and_assert_no_error(code, vec![])
-}
+use move_bytecode_verifier::verify_module;
+use move_ir_to_bytecode::{compiler::compile_module, parser::parse_module};
 
 fn compile_module_string_impl(
     code: &str,
